Please create a Social Feed page linked to the My Feed button in the banner. This should replace the current home page.
The main feature of the social feed is the scrolling block in which posts from other users are updated there. The social feed shows the user profile picture + profile name, and the body text (capped at 300 characters per post).

Posts are one of two options:

A ‘project post’; a term for an update automatically made from the linked Google Docs version history update. These can also be manually made by the user. Each ‘project post’ will be formatted by Book Title, the added or deleted word counts, and a generated summary of the changes made. A button is available in the bottom right corner that will bring the viewer to the poster’s project page so that the viewer can read a summary of the book and find more information about the project.

A ‘personal post’; a post that is familiar to social media users. This is a status update where users can share anything from what they’ve done that day or speculate about books. Essentially, a post published to the public made with a rich text editor, similar to a Facebook Status or a Tweet.

On either a Project Post or a Personal Post, users are able to click on the profile picture or user name to travel to the user’s page. There are also buttons for commenting, sharing, and liking the post, allowing other users to interact.

On the right side of the page is a WYSIWYG Rich Text Editor. Users who have signed in on their profile are able to make personal posts through this editor, published to the public after hitting the ‘publish’ button.

Underneath is a block for Google Ads integration to earn income for the site. The social feed can be changed by following users, in which followed users posts will appear at the top of the social feed. Authors who have selected interested genres will also be shown other authors who have indicated matching interested genres on the social feed. For example, an author who has selected that they are interested in Horror and Science Fiction will be shown other authors on the social feed who have indicated they also like Horror, Science Fiction, or both.