Please create a Projects page linked to the My Projects button in the banner.
Users must be able to create, edit, and view ‘projects’ as an extension of the Google Docs software. Projects are an encapsulation of written work viewed as a title, body text, and word count tracker. Updates made to the linked Google Doc will automatically update the linked project’s word count tracker as the same updates are posted to the social feed. Users must be able to edit the information of their project independently from Google Docs as well; title, summary, and word count included. This prevents irreversible errors from occurring.

Users can make up to 3 projects at any time. I’d like for a maximum of 9 projects to be possible as we will be holding the upgrade behind a paywall to encourage users to buy membership in the future.

For each project, users can edit the title (30 character limit) and write a summary (130 character limit). The status of a project can be selected from a dropdown menu: In Progress, Upcoming, or Completed.

Users can also set word count goals by clicking on the bar and updating the number values from 0. Depending on percentage of goal from 0 (e.g. 40,000/100,000 = 40%, 15,000/135,000 = 11% ) shows on progress bar as a green slider moving left to right. This number value can be manually edited at any time.

The difference between most recent update + current update (e.g. 4000 -> 4500 = +500) is the updated amount published to social feed.

For users that have linked their Google Account and selected Google Docs, this information will be automatically updated with option to edit. Counted words from linked google docs will automatically update. There is also a button option to link or re-link to Google Docs as well, in case users want to opt into the feature in the future.

Below the projects will be a contributions tracker. Like Github, the contributions tracker tracks the amount of changes made in a 24 hour window based on version history. The contributions can be tracked from the date that a profile is created until present day, with arrows allowing navigation across 90 day increments until the maximum is reached (the date of creation or present day). The total amount of changes across every project is the sum updated to the contributions tracker. If a user does not have a linked google docs file for the project, then contributions tracker is updated each time the user updates a word count on a project.

For example: If a user edits Project 1 once, Project 2 two times, and Project 3 two times within 24 hours, then there will be a total daily contribution of 5. If a user edits Project 1 once, but doesn’t edit Project 2 or Project 3, then there will be a total daily contribution of 1.

If a user manually updates Project 1 with an update of +500 words, then there will be a total daily contribution of 1. This does stack on top of linked Google Docs projects; 5 + 1 = 6 total daily contributions.

The contribution tracker days will be coded in different shades of green depending on the amount of contributions per day in increments of 5: 0 being grey, 1-5 being light green, 5-10 being green, 10-15 being dark green, and 15+ being a very dark green. Making both Projects and Contributions available to the public can be toggled off with a button, but is default on. If toggled off, users viewing the profile will not be able to see the projects and/or contribution tracker.

You can view the title, summary, word count, progress, and summary of each project on this page. The status of each project via the selected option under the title.

The scroll arrow on the right and left side allows users to navigate multiples of the projects if they have more than 3 (premium feature).

The contributions tracker shows a complete, updated history of changes made to each document through time. Linked projects with longer version histories may update the contribution tracker beyond the account creation start date - whichever has an older history.