Make the following changes:
,

Find this element:

  <ProfilePage>

with the text content:

    John DoeAspiring fiction author.123 Followers456 FollowingFantasySci-FiHorrorMysteryThrillerRomanceHistorical FictionBiographyProjectsSocial FeedAbout MeA little bit about myself.Aspiring fiction author.Powered by Firebase Studio

Relevant files:

    - https://6000-idx-studio-1744498864021.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-idx-studio-1744498864021.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Please move the follower and following count to be right-aligned in the cover photo's box with a slightly translucent grey box behind it. The followers and following counts, when clicked on, should list the users who follow or are followed by the user.